Fix User model: email unique validator, date. Add controller floder and create db.js
