+++ package.json
@@ -11,6 +11,7 @@
     "helmet": "^4.4.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.12.0",
+    "mongoose-unique-validator": "^2.0.3",
     "validator": "^13.5.2"
   },
   "devDependencies": {
